Made login
You can login with email and password and with google. You can also register. The only thing with register is that you have to fill in like you are about to login, but instead of login you click on register. Thats the only thing but it works. Now I just need to start with the profile page